pyproject.toml
[project]
name = "tether"
version = "0.1.0"
description = "Lightweight JSON-RPC client for Arduino over serial"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Revise Robotics" },
]
readme = "README.md"
keywords = ["arduino", "serial", "rpc", "json", "iot", "embedded"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: System :: Hardware",
]
dependencies = [
    "pyserial>=3.5",
]

[project.urls]
Homepage = "https://github.com/Revise-Robotics/tether-python"
Repository = "https://github.com/Revise-Robotics/tether-python"
Issues = "https://github.com/Revise-Robotics/tether-python/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
